Add files via upload
This work involves the creation of a YouTube video downloader using the Pytube library and a graphical user interface (GUI) using the Tkinter library in Python. The application allows users to enter the URL of a YouTube video and download it to their computer in a selected video quality. The GUI provides a simple and user-friendly interface for users to input the video URL and select the desired quality. The application uses Pytube to handle the video downloading process and Tkinter to create the GUI.

While downloading YouTube videos is generally prohibited by YouTube's terms of service, this work is presented as an educational and technical example of how to use Python libraries to create a useful application. It is important for users to understand the potential legal consequences of downloading YouTube videos without permission and to use this application for personal, non-commercial purposes only.

Dependencies : pip install pytube , pip install tkinter